FOVThingDashboard/app/requirements.txt
PyYAML==6.0.2
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchfiles==0.24.0
websockets==13.0.1
psutil==5.9.8
fastapi==0.114.0
starlette==0.38.5
uvicorn==0.30.6
PyJWT==2.9.0
passlib[bcrypt]>=1.7.4
awscrt==0.21.1
awsiotsdk==1.22.0
